requirements.txt
pandas
numpy
requests
python-dotenv
scikit-learn
joblib
streamlit
shap
matplotlib
